Fixed problems with algorithm for 3D pen tool
Mostly, the set of edges being passed as input wasn't consistent between
the two planes, which led to nonsensical edges in some cases.
Mostly stemmed from fact that we used to create a dummy edge for first
point of pen tool.
So that's fixed now.
Some problems remain with 3d pen tool:
[ ] - plane/edge gaps still exist - need to go over this algorithm
[ ] - the algorithm itself is functioning as intended, but it might not
be expected behaviour