Added input field to inspector, fixed weird transparency issues
There’s this weirdness with rendering semi-transparent sections
apparently because of the way that z-buffers work. Basically, it was
causing the inspector to appear as though it was behind the guide plane
even though it wasn’t, which was pretty annoying
There were a bunch of suggestions, but what I ended up doing was making
a new UI sorting layer, and then setting the inspector to that layer.