REFACTOR: Moved pen mode logic into Shape data structure
THIS MIGHT HAVE BROKEN A LOT OF SHIT
This was a lot of code shifting around, and while I think it’s right,
potential that it broke something is high.
Basic idea is to have a data structure where you can just input a new
vertex, and receive a list of points that have been set - the setting
of blocks will happen in here.
Have to extend for free paint mode
This will form the base of saving for objects, selection tool,
everything really